# Conflicts:
#	src/com/company/Computer.java
#	src/com/company/Electronics.java
#	src/com/company/Main.java
#	src/com/company/Product.java
#	src/com/company/RandomAccessMemory.java
#	src/com/company/Smartphone.java
#	src/com/company/test/Test.java

+++ src/com/company/Electronics.java
@@ -1,7 +1,9 @@
 package com.company;
 
+import java.math.BigDecimal;
+
 public class Electronics extends Product {
-    public Electronics( String name, double price, int stock) {
+    public Electronics( String name, BigDecimal price, int stock) {
         super( name, price, stock);
     }
